Added unit tests, renamed objects and added checks for overwrites

+++ taskcat/project_config/config.py
@@ -13,13 +13,15 @@ from taskcat.project_config.tools import _add_parameter_values, _get_parameter_s
 yaml = YAML()
 yaml.preserve_quotes = True
 LOG = logging.getLogger(__name__)
+LOG_CONFIG = logging.getLogger("taskcat._config")
+LOG_CONFIG.setLevel(logging.ERROR)
 
 
-class TaskCatConfigGenerator:
+class ConfigGenerator:
     def __init__(
             self, main_template: str, output_file: str,
             project_root_path: str, owner_email: str,
-            aws_region: str):
+            aws_region: str, replace: bool):
         self.output_file = output_file
         self.main_template = main_template
         self.project_root_path = project_root_path
@@ -27,17 +29,26 @@ class TaskCatConfigGenerator:
         self.aws_region = aws_region
         self.repo = git.Repo(project_root_path, search_parent_directories=True)
         self.repo_name = self.repo.remotes.origin.url.split('.git')[0].split('/')[-1]
-
+        self.replace = replace
 
     def generate_config(self):
         base_path = Path(self.project_root_path).resolve()
         template_path = base_path / self.main_template
-        project_config_path = base_path / "./.taskcat.yml"
+        project_config_path = base_path / self.output_file
         global_config_path = base_path / "./.taskcat_global.yml"
+
+        # Check that the provided paths are valid and return an error if they are not
+        if not base_path.exists():
+            raise TaskCatException(f"Base path {base_path} does not exist")
+        if not template_path.exists():
+            raise TaskCatException(f"Template file {template_path} does not exist")
+
 
         LOG.warning("This is an ALPHA feature. Use with caution")
         # Read Yaml file
-        cfn = yaml.load(open(template_path, 'r', encoding="utf-8"))
+        with open(template_path, 'r', encoding="utf-8") as f:
+            cfn = yaml.load(f)
+        f.close()
         # Container for each parameter object
         parameters = {}
         # Get data for each parameter
@@ -59,10 +70,24 @@ class TaskCatConfigGenerator:
             })
         })
 
-        # Remove empty config file
-        if os.stat(project_config_path).st_size == 0:
-            os.remove(project_config_path)
-
+        # Config File Handling
+        config_file = Path(project_config_path)
+        if config_file.exists():
+            # Check if file is empty
+            if os.stat(project_config_path).st_size == 0:
+                LOG.warning("Empty project config found. "
+                            "Overwriting...")
+                os.remove(project_config_path)
+            elif os.stat(project_config_path).st_size > 0 and not self.replace:
+                raise TaskCatException("Project config file already exists. "
+                                       "No changes will be made. "
+                                       "to override use --replace flag")
+                # LOG.warning("Project config file already exists. "
+                #             "No changes will be made. to override use --replace flag")
+                # sys.exit(1)
+            else:
+                LOG.warning("Project config file already exists. "
+                            "Overwriting existing file")
         try:
             config = Config.create(
                 args=cfg_dict,
@@ -74,6 +99,7 @@ class TaskCatConfigGenerator:
                     "TASKCAT_PROJECT_SHORTEN_STACK_NAME": "True",
                     "TASKCAT_PROJECT_S3_REGIONAL_BUCKETS": "True"},
             )
+            LOG.info("Config file created successfully")
         except TaskCatException as e:
             LOG.error(e)
             sys.exit(1)
@@ -84,4 +110,4 @@ class TaskCatConfigGenerator:
             config_dict.pop('general')
             yaml.dump(config_dict, outfile)
         outfile.close()
-        print(_get_parameter_stats(cfg_dict['project']['parameters']))
+        LOG.info(_get_parameter_stats(cfg_dict['project']['parameters']))
